Revert "Plan 9 from Bell Labs Fourth Edition 2013-09-17"
This reverts commit 6ca6a3e703ee2ec4aed99c2177f71d7f127da6d9.

+++ sys/src/ape/lib/bsd/ntohl.c
@@ -1,5 +1,5 @@
 unsigned long
-ntohl(unsigned long x)
+ntohl(int x)
 {
 	unsigned long n;
 	unsigned char *p;
@@ -24,7 +24,7 @@ htonl(unsigned long h)
 }
 
 unsigned short
-ntohs(unsigned short x)
+ntohs(int x)
 {
 	unsigned short n;
 	unsigned char *p;
